third_party/blink/web_tests/external/wpt/editing/other/insertparagraph-in-child-of-head.tentative_designMode=off_white-space=pre-line-expected.txt

This is a testharness.js-based test.
[FAIL] insertParagraph in empty <title style="display:block;white-space:pre-line"> should not split it
  assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>\\n\\n</title></head><body><br></body>" but got "<head><title></title></head><body><br><div><br></div></body>"
[FAIL] insertParagraph in <title style="display:block;white-space:pre-line"> containing text should not split it
  assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>ab\\ncd</title></head><body><br></body>" but got "<head><title>ab</title><title style=\\"display:block;white-space:pre-line\\">cd</title></head><body><br></body>"
[FAIL] insertParagraph in empty <style style="display:block;white-space:pre-line"> should not split it
  assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>iframe</title><style>\\n\\n</style></head><body><br></body>" but got "<head><title>iframe</title><style></style></head><body><br><div><br></div></body>"
[FAIL] insertParagraph in <style style="display:block;white-space:pre-line"> containing text should not split it
  assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>iframe</title><style>ab\\ncd</style></head><body><br></body>" but got "<head><title>iframe</title><style>ab</style><style style=\\"display:block;white-space:pre-line\\">cd</style></head><body><br></body>"
[FAIL] insertParagraph in empty <script style="display:block;white-space:pre-line"> should not split it
  assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>iframe</title><script>\\n\\n</script></head><body><br></body>" but got "<head><title>iframe</title><script></script></head><body><br><div><br></div></body>"
[FAIL] insertParagraph in <script style="display:block;white-space:pre-line"> containing text should not split it
  assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>iframe</title><script>// ab\\n// cd</script></head><body><br></body>" but got "<head><title>iframe</title><script>// ab</script><script style=\\"display:block;white-space:pre-line\\">// cd</script></head><body><br></body>"
[FAIL] insertParagraph in empty <div style="white-space:pre-line"> in the <head> should split the <div>
  assert_in_array: The <div> should be split value "<head><title>iframe</title><div style=\\"white-space:pre-line\\"></div></head><body><br><div><br></div></body>" not in array ["<head><title>iframe</title><div style=\\"white-space:pre-line\\">\\n</div><div style=\\"white-space:pre-line\\">\\n</div></head><body><br></body>", "<head><title>iframe</title><div style=\\"white-space:pre-line\\"><br></div><div style=\\"white-space:pre-line\\"><br></div></head><body><br></body>"]
Harness: the test ran to completion.
